requirements.txt
PyQt6>=6.4.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6
numpy>=1.21.0
Pillow>=9.0.0
requests>=2.28.0